requirements.txt
Django
requests>=2.28
celery>=5.2
redis>=4.3
psycopg2-binary>=2.9
python-dotenv>=0.21
reportlab>=4.0
pytest
djangorestframework